Create json_ key_finder_console.py
Loads a file called object.json, then attempts to fix any JSON errors in the file. Takes input of a key you're looking for and then searches the JSON for all instances of the requested key. Outputs the path in the JSON in python format of each instance.

Ex:
Enter the key to find: userName

Paths to 'userName':
event['detail']['userIdentity']['userName']
event['userName']